requirements.txt
streamlit
openai
google-genai
pandas
pycountry
Pillow
torch
transformers
accelerate
